Updated namespace Added EBD_mutation.R EBD_parent_sampler now has separate subfunction for random selection. Faster performance EBD_pop now takes phenotypes as integer instead of numeric

+++ R/EBD_pop.R
@@ -77,7 +77,7 @@ EBD.insert_genotype = function( pop, genotypes, index ){
     pop@phenotype$phenotype[ index ] = bin2int( genotypes )
 }
 
-setMethod( "pop.set", signature( pop = "EBD_pop",  phenotypes = "numeric", index = "integer", genotypes = "missing" ), function( pop, phenotypes, index ){
+setMethod( "pop.set", signature( pop = "EBD_pop",  phenotypes = "integer", index = "integer", genotypes = "missing" ), function( pop, phenotypes, index ){
     EBD.insert_phenotype( pop, phenotypes, index )
 } )
